+++ etc/telegraf.conf
@@ -1377,8 +1377,56 @@
 #   servers = ["http://localhost:8098"]
 
 
-# # Reads oids value from one or many snmp agents
+# # Retrieves SNMP values from remote agents
 # [[inputs.snmp]]
+#   agents = [ "127.0.0.1:161" ]
+#   version = 2 # Values: 1, 2, or 3
+#
+#   ## SNMPv1 & SNMPv2 parameters
+#   community = "public"
+#
+#   ## SNMPv2 & SNMPv3 parameters
+#   max_repetitions = 50
+#
+#   ## SNMPv3 parameters
+#   #sec_name = "myuser"
+#   #auth_protocol = "md5"         # Values: "MD5", "SHA", ""
+#   #auth_password = "password123"
+#   #sec_level = "authNoPriv"      # Values: "noAuthNoPriv", "authNoPriv", "authPriv"
+#   #context_name = ""
+#   #priv_protocol = ""            # Values: "DES", "AES", ""
+#   #priv_password = ""
+#
+#   ## measurement name
+#   name = "system"
+#   [[inputs.snmp.field]]
+#     name = "hostname"
+#     oid = ".1.2.3.0.1.1"
+#   [[inputs.snmp.field]]
+#     name = "uptime"
+#     oid = ".1.2.3.0.1.200"
+#   [[inputs.snmp.field]]
+#     name = "load"
+#     oid = ".1.2.3.0.1.201"
+#
+#   [[inputs.snmp.table]]
+#     # measurement name
+#     name = "remote_servers"
+#     inherit_tags = [ "hostname" ]
+#     [[inputs.snmp.table.field]]
+#       name = "server"
+#       oid = ".1.2.3.0.0.0"
+#       is_tag = true
+#     [[inputs.snmp.table.field]]
+#       name = "connections"
+#       oid = ".1.2.3.0.0.1"
+#     [[inputs.snmp.table.field]]
+#       name = "latency"
+#       oid = ".1.2.3.0.0.2"
+
+
+# # DEPRECATED: This will be removed in a future release.
+# [[inputs.snmp_legacy]]
 #   ## Use 'oids.txt' file to translate oids to names
 #   ## To generate 'oids.txt' you need to run:
 #   ##   snmptranslate -m all -Tz -On | sed -e 's/"//g' > /tmp/oids.txt
